+++ scraper.py
@@ -26,6 +26,7 @@ from psycopg2 import sql
 import pandas as pd
 from jobspy import scrape_jobs
 
+from database import DatabaseConfig, JobDatabase
 import warnings
 # Suppress pandas SQLAlchemy warnings for psycopg2 connections
 warnings.filterwarnings('ignore', message='pandas only supports SQLAlchemy connectable')
@@ -55,938 +56,6 @@ logging.basicConfig(
 logger = logging.getLogger(__name__)
 
 
-class DatabaseConfig:
-    """Class to handle database configuration loading."""
-    
-    def __init__(self, config_path: str = None, 
-                 database_type: str = "production"):
-        """Initialize with configuration file path and database type.
-        
-        Args:
-            config_path: Path to the database configuration file
-            database_type: Either 'production' or 'working' to select database config
-        """
-        if config_path is None:
-            config_path = os.path.join(SCRIPT_DIR, "configs", "db", "db_config.json")
-        self.config_path = config_path
-        self.database_type = database_type
-        self.config = self._load_config()
-        
-    def _load_config(self) -> Dict:
-        """Load database configuration from file.
-        
-        Returns:
-            Dictionary containing database configuration
-        """
-        if not os.path.exists(self.config_path):
-            raise FileNotFoundError(f"Database config file not found: {self.config_path}")
-            
-        with open(self.config_path, 'r') as f:
-            return json.load(f)
-    
-    def get_connection_params(self) -> Dict:
-        """Get database connection parameters.
-        
-        Returns:
-            Dictionary of connection parameters for psycopg2
-        """
-        # Determine which database config to use
-        if self.database_type == "working" and 'working_database' in self.config:
-            db_config = self.config['working_database']
-        elif self.database_type == "production" and 'production_database' in self.config:
-            db_config = self.config['production_database']
-        elif 'database' in self.config:
-            # Fallback to old format for backwards compatibility
-            db_config = self.config['database']
-        else:
-            raise KeyError(f"No database configuration found for type: {self.database_type}")
-            
-        conn_config = self.config.get('connection', {})
-        
-        return {
-            'host': db_config['host'],
-            'port': db_config['port'],
-            'database': db_config['database'],
-            'user': db_config['username'],
-            'password': db_config['password'],
-            'connect_timeout': conn_config.get('connect_timeout', 30),
-            'application_name': 'jobscraps_scraper'
-        }
-    
-    def get_retry_config(self) -> Dict:
-        """Get retry configuration.
-        
-        Returns:
-            Dictionary with retry settings
-        """
-        return self.config.get('connection', {})
-
-
-class JobDatabase:
-    """Class to handle database operations for job data."""
-    
-    def __init__(self, config_path: str = None,
-                 database_type: str = "production"):
-        """Initialize the database connection.
-        
-        Args:
-            config_path: Path to the database configuration file
-            database_type: Either 'production' or 'working' to select database config
-        """
-        if config_path is None:
-            config_path = os.path.join(SCRIPT_DIR, "configs", "db", "db_config.json")
-        self.db_config = DatabaseConfig(config_path, database_type)
-        self.database_type = database_type
-        self.conn = None
-        self._connect_with_retry()
-        self.create_tables()
-        logger.info(f"PostgreSQL database initialized ({database_type})")
-        
-    def _connect_with_retry(self) -> None:
-        """Connect to database with retry logic."""
-        retry_config = self.db_config.get_retry_config()
-        max_attempts = retry_config.get('retry_attempts', 3)
-        retry_delay = retry_config.get('retry_delay', 5)
-        
-        for attempt in range(max_attempts):
-            try:
-                conn_params = self.db_config.get_connection_params()
-                self.conn = psycopg2.connect(**conn_params)
-                self.conn.autocommit = False
-                logger.info(f"Connected to PostgreSQL database successfully")
-                return
-            except psycopg2.Error as e:
-                logger.warning(f"Database connection attempt {attempt + 1} failed: {e}")
-                if attempt < max_attempts - 1:
-                    time.sleep(retry_delay)
-                else:
-                    raise psycopg2.Error(f"Failed to connect after {max_attempts} attempts: {e}")
-    
-    def _ensure_connection(self) -> None:
-        """Ensure database connection is active, reconnect if needed."""
-        try:
-            if self.conn.closed:
-                logger.warning("Database connection is closed, reconnecting...")
-                self._connect_with_retry()
-        except (psycopg2.Error, AttributeError):
-            logger.warning("Database connection error, reconnecting...")
-            self._connect_with_retry()
-
-    def create_backup(self, backup_type: str = 'auto', reason: str = '') -> Dict:
-        """Create a PostgreSQL backup using pg_dump.
-        
-        Args:
-            backup_type: Type of backup ('auto' or 'manual')
-            reason: Reason for backup (e.g., 'scraping', 'deletion', 'working_copy')
-            
-        Returns:
-            Dictionary with backup information
-        """
-        try:
-            # Create backup directory
-            backup_dir = os.path.join(SCRIPT_DIR, "backups", "DatabaseBackups")
-            os.makedirs(backup_dir, exist_ok=True)
-            
-            # Get connection parameters
-            conn_params = self.db_config.get_connection_params()
-            
-            # Create filename with timestamp and reason
-            timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-            reason_suffix = f"_{reason}" if reason else ""
-            filename = f"jobscraps_{timestamp}_{backup_type}{reason_suffix}.sql.gz"
-            backup_path = os.path.join(backup_dir, filename)
-            
-            # Build pg_dump command
-            cmd = [
-                'pg_dump',
-                '-h', conn_params['host'],
-                '-p', str(conn_params['port']),
-                '-U', conn_params['user'],
-                '-d', conn_params['database'],
-                '--compress=9',
-                '--verbose',
-                '--file', backup_path
-            ]
-            
-            # Set password environment variable
-            env = os.environ.copy()
-            env['PGPASSWORD'] = conn_params['password']
-            
-            logger.info(f"Creating backup: {filename}")
-            start_time = time.time()
-            
-            # Execute pg_dump with retry logic
-            max_attempts = 3
-            for attempt in range(max_attempts):
-                try:
-                    result = subprocess.run(
-                        cmd, 
-                        env=env, 
-                        capture_output=True, 
-                        text=True, 
-                        timeout=300  # 5 minute timeout
-                    )
-                    
-                    if result.returncode == 0:
-                        backup_time = time.time() - start_time
-                        file_size = os.path.getsize(backup_path)
-                        file_size_mb = file_size / (1024 * 1024)
-                        
-                        backup_info = {
-                            'filename': filename,
-                            'path': backup_path,
-                            'size_bytes': file_size,
-                            'size_mb': round(file_size_mb, 1),
-                            'duration_seconds': round(backup_time, 1),
-                            'timestamp': timestamp,
-                            'reason': reason,
-                            'backup_type': backup_type
-                        }
-                        
-                        # Update backup manifest
-                        self._update_backup_manifest(backup_info)
-                        
-                        logger.info(f"Backup created successfully: {filename} ({file_size_mb:.1f} MB in {backup_time:.1f}s)")
-                        return backup_info
-                        
-                    else:
-                        error_msg = result.stderr or result.stdout
-                        logger.warning(f"pg_dump attempt {attempt + 1} failed: {error_msg}")
-                        if attempt < max_attempts - 1:
-                            time.sleep(5)  # Wait before retry
-                        
-                except subprocess.TimeoutExpired:
-                    logger.warning(f"pg_dump attempt {attempt + 1} timed out")
-                    if attempt < max_attempts - 1:
-                        time.sleep(5)
-                except Exception as e:
-                    logger.warning(f"pg_dump attempt {attempt + 1} error: {e}")
-                    if attempt < max_attempts - 1:
-                        time.sleep(5)
-            
-            # If all attempts failed
-            if os.path.exists(backup_path):
-                os.remove(backup_path)  # Clean up partial file
-            raise Exception(f"pg_dump failed after {max_attempts} attempts")
-            
-        except Exception as e:
-            logger.error(f"Backup creation failed: {e}")
-            raise
-    
-    def _update_backup_manifest(self, backup_info: Dict) -> None:
-        """Update the backup manifest file.
-        
-        Args:
-            backup_info: Dictionary containing backup information
-        """
-        manifest_path = os.path.join(SCRIPT_DIR, "backups", "DatabaseBackups", "backup_manifest.json")
-        
-        try:
-            # Load existing manifest
-            if os.path.exists(manifest_path):
-                with open(manifest_path, 'r') as f:
-                    manifest = json.load(f)
-            else:
-                manifest = {"backups": []}
-            
-            # Add new backup info
-            manifest["backups"].append(backup_info)
-            
-            # Update summary statistics
-            manifest["total_backups"] = len(manifest["backups"])
-            total_size = sum(backup['size_bytes'] for backup in manifest["backups"])
-            manifest["total_size_gb"] = round(total_size / (1024**3), 2)
-            
-            if manifest["backups"]:
-                manifest["oldest_backup"] = min(backup['timestamp'] for backup in manifest["backups"])
-                manifest["newest_backup"] = max(backup['timestamp'] for backup in manifest["backups"])
-                manifest["last_updated"] = datetime.now().strftime("%Y%m%d_%H%M%S")
-            
-            # Save updated manifest
-            with open(manifest_path, 'w') as f:
-                json.dump(manifest, f, indent=2)
-                
-        except Exception as e:
-            logger.warning(f"Failed to update backup manifest: {e}")
-    
-    def manage_backup_retention(self) -> Dict:
-        """Manage backup retention policy.
-        
-        Returns:
-            Dictionary with retention summary
-        """
-        backup_dir = os.path.join(SCRIPT_DIR, "backups", "DatabaseBackups")
-        manifest_path = os.path.join(backup_dir, "backup_manifest.json")
-        
-        try:
-            if not os.path.exists(manifest_path):
-                return {"action": "no_manifest", "message": "No backup manifest found"}
-            
-            with open(manifest_path, 'r') as f:
-                manifest = json.load(f)
-            
-            backups = manifest.get("backups", [])
-            total_size_gb = manifest.get("total_size_gb", 0)
-            
-            # Check if cleanup needed
-            needs_cleanup = len(backups) > 48 or total_size_gb > 4.8
-            
-            if not needs_cleanup:
-                return {
-                    "action": "no_cleanup_needed",
-                    "total_backups": len(backups),
-                    "total_size_gb": total_size_gb
-                }
-            
-            # Sort backups by timestamp (oldest first)
-            backups.sort(key=lambda x: x['timestamp'])
-            
-            # Calculate how many to remove
-            target_count = 40
-            to_remove = max(0, len(backups) - target_count)
-            
-            if total_size_gb > 4.8:
-                # Remove additional backups to get under size limit
-                size_removed = 0
-                while len(backups) > to_remove and total_size_gb > 4.5:
-                    oldest_backup = backups[to_remove]
-                    size_removed += oldest_backup['size_bytes']
-                    total_size_gb -= oldest_backup['size_bytes'] / (1024**3)
-                    to_remove += 1
-            
-            # Remove old backup files
-            removed_backups = []
-            for i in range(to_remove):
-                backup = backups[i]
-                backup_path = backup['path']
-                
-                if os.path.exists(backup_path):
-                    os.remove(backup_path)
-                    removed_backups.append(backup['filename'])
-                    logger.info(f"Removed old backup: {backup['filename']}")
-            
-            # Update manifest
-            manifest["backups"] = backups[to_remove:]
-            manifest["total_backups"] = len(manifest["backups"])
-            
-            if manifest["backups"]:
-                total_size = sum(backup['size_bytes'] for backup in manifest["backups"])
-                manifest["total_size_gb"] = round(total_size / (1024**3), 2)
-                manifest["oldest_backup"] = min(backup['timestamp'] for backup in manifest["backups"])
-                manifest["last_cleanup"] = datetime.now().strftime("%Y%m%d_%H%M%S")
-            else:
-                manifest["total_size_gb"] = 0
-            
-            with open(manifest_path, 'w') as f:
-                json.dump(manifest, f, indent=2)
-            
-            return {
-                "action": "cleanup_performed",
-                "removed_count": len(removed_backups),
-                "removed_files": removed_backups,
-                "remaining_backups": len(manifest["backups"]),
-                "total_size_gb": manifest["total_size_gb"]
-            }
-            
-        except Exception as e:
-            logger.error(f"Backup retention management failed: {e}")
-            return {"action": "error", "message": str(e)}
-    
-    def list_backups(self) -> List[Dict]:
-        """List available backups.
-        
-        Returns:
-            List of backup information dictionaries
-        """
-        manifest_path = os.path.join(SCRIPT_DIR, "backups", "DatabaseBackups", "backup_manifest.json")
-        
-        try:
-            if not os.path.exists(manifest_path):
-                return []
-            
-            with open(manifest_path, 'r') as f:
-                manifest = json.load(f)
-            
-            backups = manifest.get("backups", [])
-            # Sort by timestamp (newest first)
-            backups.sort(key=lambda x: x['timestamp'], reverse=True)
-            
-            return backups
-            
-        except Exception as e:
-            logger.error(f"Failed to list backups: {e}")
-            return []
-    
-    def restore_backup(self, backup_filename: str) -> bool:
-        """Restore database from backup file.
-        
-        Args:
-            backup_filename: Name of the backup file to restore
-            
-        Returns:
-            True if restore was successful, False otherwise
-        """
-        backup_path = os.path.join(SCRIPT_DIR, "backups", "DatabaseBackups", backup_filename)
-        
-        if not os.path.exists(backup_path):
-            logger.error(f"Backup file not found: {backup_path}")
-            return False
-        
-        try:
-            conn_params = self.db_config.get_connection_params()
-            
-            # Close current connection
-            if self.conn and not self.conn.closed:
-                self.conn.close()
-            
-            # Build psql restore command
-            cmd = [
-                'psql',
-                '-h', conn_params['host'],
-                '-p', str(conn_params['port']),
-                '-U', conn_params['user'],
-                '-d', conn_params['database'],
-                '-f', backup_path,
-                '--quiet'
-            ]
-            
-            env = os.environ.copy()
-            env['PGPASSWORD'] = conn_params['password']
-            
-            logger.info(f"Restoring from backup: {backup_filename}")
-            result = subprocess.run(cmd, env=env, capture_output=True, text=True, timeout=600)
-            
-            if result.returncode == 0:
-                logger.info(f"Database restored successfully from {backup_filename}")
-                # Reconnect after restore
-                self._connect_with_retry()
-                return True
-            else:
-                logger.error(f"Restore failed: {result.stderr}")
-                return False
-                
-        except Exception as e:
-            logger.error(f"Restore operation failed: {e}")
-            return False
-    
-    def test_backup(self, backup_filename: str) -> bool:
-        """Test backup file integrity.
-        
-        Args:
-            backup_filename: Name of the backup file to test
-            
-        Returns:
-            True if backup is valid, False otherwise
-        """
-        backup_path = os.path.join(SCRIPT_DIR, "backups", "DatabaseBackups", backup_filename)
-        
-        if not os.path.exists(backup_path):
-            logger.error(f"Backup file not found: {backup_path}")
-            return False
-        
-        try:
-            # Test if file can be read and is valid gzip
-            if backup_filename.endswith('.gz'):
-                import gzip
-                with gzip.open(backup_path, 'rt') as f:
-                    # Read first few lines to verify it's a valid SQL dump
-                    lines = [f.readline() for _ in range(10)]
-                    
-                # Check for PostgreSQL dump headers
-                content = ''.join(lines)
-                if 'PostgreSQL database dump' in content or 'CREATE TABLE' in content:
-                    logger.info(f"Backup file {backup_filename} appears to be valid")
-                    return True
-                else:
-                    logger.error(f"Backup file {backup_filename} does not appear to be a valid PostgreSQL dump")
-                    return False
-            else:
-                logger.error(f"Backup file {backup_filename} is not a compressed file")
-                return False
-                
-        except Exception as e:
-            logger.error(f"Backup test failed: {e}")
-            return False
-
-    def create_tables(self) -> None:
-        """Create necessary tables if they don't exist."""
-        self._ensure_connection()
-        
-        with self.conn.cursor() as cursor:
-            # Create scraped_jobs table
-            cursor.execute('''
-            CREATE TABLE IF NOT EXISTS scraped_jobs (
-                id TEXT PRIMARY KEY,
-                site TEXT,
-                job_url TEXT,
-                job_url_direct TEXT,
-                title TEXT,
-                company TEXT,
-                location TEXT,
-                date_posted TEXT,
-                job_type TEXT,
-                salary_source TEXT,
-                interval TEXT,
-                min_amount DECIMAL(12,2),
-                max_amount DECIMAL(12,2),
-                currency TEXT,
-                is_remote BOOLEAN,
-                job_level TEXT,
-                job_function TEXT,
-                listing_type TEXT,
-                emails TEXT,
-                description TEXT,
-                company_industry TEXT,
-                company_url TEXT,
-                company_logo TEXT,
-                company_url_direct TEXT,
-                company_addresses TEXT,
-                company_num_employees TEXT,
-                company_revenue TEXT,
-                company_description TEXT,
-                skills TEXT,
-                experience_range TEXT,
-                company_rating TEXT,
-                company_reviews_count TEXT,
-                vacancy_count TEXT,
-                work_from_home_type TEXT,
-                date_scraped TIMESTAMP,
-                search_query TEXT
-            )
-            ''')
-            
-            # Create search_history table
-            cursor.execute('''
-            CREATE TABLE IF NOT EXISTS search_history (
-                id SERIAL PRIMARY KEY,
-                search_query TEXT,
-                parameters TEXT,
-                timestamp TIMESTAMP,
-                jobs_found INTEGER
-            )
-            ''')
-            
-            # Ensure pg_trgm extension exists (for text similarity and pattern matching)
-            cursor.execute("CREATE EXTENSION IF NOT EXISTS pg_trgm")
-            
-            # Create indexes - Updated to match schema and optimize for degree requirement queries
-            index_queries = [
-                # Basic field indexes
-                "CREATE INDEX IF NOT EXISTS idx_scraped_jobs_title ON scraped_jobs(title)",
-                "CREATE INDEX IF NOT EXISTS idx_scraped_jobs_company ON scraped_jobs(company)",
-                "CREATE INDEX IF NOT EXISTS idx_scraped_jobs_location ON scraped_jobs(location)",
-                "CREATE INDEX IF NOT EXISTS idx_scraped_jobs_site ON scraped_jobs(site)",
-                "CREATE INDEX IF NOT EXISTS idx_scraped_jobs_is_remote ON scraped_jobs(is_remote)",
-                "CREATE INDEX IF NOT EXISTS idx_scraped_jobs_date_posted ON scraped_jobs(date_posted)",
-                "CREATE INDEX IF NOT EXISTS idx_scraped_jobs_date_scraped ON scraped_jobs(date_scraped)",
-                
-                # Search query indexes
-                "CREATE INDEX IF NOT EXISTS idx_scraped_jobs_search_query ON scraped_jobs(search_query)",
-                "CREATE INDEX IF NOT EXISTS idx_scraped_jobs_search_query_lower ON scraped_jobs(LOWER(search_query))",
-                
-                # Composite indexes for common query patterns
-                "CREATE INDEX IF NOT EXISTS idx_scraped_jobs_title_company ON scraped_jobs(title, company)",
-                "CREATE INDEX IF NOT EXISTS idx_scraped_jobs_company_location ON scraped_jobs(company, location)",
-                
-                # TEXT SEARCH OPTIMIZATION - Critical for degree requirement queries
-                # GIN index with pg_trgm for efficient ILIKE pattern matching on description
-                "CREATE INDEX IF NOT EXISTS idx_scraped_jobs_description_gin ON scraped_jobs USING gin (description gin_trgm_ops)",
-                
-                # Search history indexes
-                "CREATE INDEX IF NOT EXISTS idx_search_history_search_query ON search_history(search_query)",
-                "CREATE INDEX IF NOT EXISTS idx_search_history_timestamp ON search_history(timestamp)"
-            ]
-            
-            for query in index_queries:
-                try:
-                    cursor.execute(query)
-                    logger.debug(f"Successfully created/verified index: {query.split('idx_')[1].split(' ')[0] if 'idx_' in query else 'extension'}")
-                except Exception as e:
-                    logger.warning(f"Index creation warning: {e}")
-                    # Continue with other indexes even if one fails
-            
-            self.conn.commit()
-            logger.info("Database tables and indexes created/verified successfully")
-        
-    def insert_jobs(self, jobs_df: pd.DataFrame, search_query: str) -> int:
-        """Insert new jobs into the database, maintaining uniqueness by ID.
-        
-        Args:
-            jobs_df: DataFrame containing job data
-            search_query: The search query used to find these jobs
-            
-        Returns:
-            Number of new jobs inserted
-        """
-        self._ensure_connection()
-        
-        # Add date_scraped and search_query columns
-        jobs_df['date_scraped'] = datetime.now()
-        jobs_df['search_query'] = search_query
-        
-        # Get existing job IDs to check for duplicates
-        with self.conn.cursor() as cursor:
-            cursor.execute("SELECT id FROM scraped_jobs")
-            existing_ids = {row[0] for row in cursor.fetchall()}
-        
-        # Filter out jobs that already exist in the database
-        if 'id' in jobs_df.columns:
-            new_jobs_df = jobs_df[~jobs_df['id'].isin(existing_ids)]
-        else:
-            # If no ID column, create one
-            jobs_df['id'] = jobs_df.apply(
-                lambda row: f"{row.get('site', 'unknown')}_{row.get('job_url', '')[-20:]}",
-                axis=1
-            )
-            new_jobs_df = jobs_df[~jobs_df['id'].isin(existing_ids)]
-        
-        # Convert boolean columns properly for PostgreSQL
-        if 'is_remote' in new_jobs_df.columns:
-            new_jobs_df['is_remote'] = new_jobs_df['is_remote'].astype(bool)
-        
-        # Insert new jobs
-        new_jobs_count = len(new_jobs_df)
-        if new_jobs_count > 0:
-            # Convert DataFrame to list of tuples for batch insert
-            columns = [
-                'id', 'site', 'job_url', 'job_url_direct', 'title', 'company', 'location',
-                'date_posted', 'job_type', 'salary_source', 'interval', 'min_amount', 'max_amount',
-                'currency', 'is_remote', 'job_level', 'job_function', 'listing_type', 'emails',
-                'description', 'company_industry', 'company_url', 'company_logo', 'company_url_direct',
-                'company_addresses', 'company_num_employees', 'company_revenue', 'company_description',
-                'skills', 'experience_range', 'company_rating', 'company_reviews_count', 'vacancy_count',
-                'work_from_home_type', 'date_scraped', 'search_query'
-            ]
-            
-            # Ensure all columns exist in DataFrame
-            for col in columns:
-                if col not in new_jobs_df.columns:
-                    new_jobs_df[col] = None
-            
-            # Prepare data for insertion
-            data_to_insert = []
-            for _, row in new_jobs_df.iterrows():
-                row_data = []
-                for col in columns:
-                    value = row[col]
-                    # Handle NaN values
-                    if pd.isna(value):
-                        row_data.append(None)
-                    else:
-                        row_data.append(value)
-                data_to_insert.append(tuple(row_data))
-            
-            # Build insert query
-            insert_query = sql.SQL(
-                "INSERT INTO scraped_jobs ({}) VALUES ({})"
-            ).format(
-                sql.SQL(', ').join(map(sql.Identifier, columns)),
-                sql.SQL(', ').join(sql.Placeholder() * len(columns))
-            )
-            
-            with self.conn.cursor() as cursor:
-                psycopg2.extras.execute_batch(cursor, insert_query, data_to_insert)
-                self.conn.commit()
-                
-            logger.info(f"Inserted {new_jobs_count} new jobs into database")
-        else:
-            logger.info("No new jobs to insert")
-            
-        return new_jobs_count
-    
-    def log_search(self, search_query: str, parameters: Dict, jobs_found: int) -> None:
-        """Log search history.
-        
-        Args:
-            search_query: The search query used
-            parameters: The parameters used for the search
-            jobs_found: Number of jobs found in this search
-        """
-        self._ensure_connection()
-        
-        with self.conn.cursor() as cursor:
-            cursor.execute(
-                "INSERT INTO search_history (search_query, parameters, timestamp, jobs_found) VALUES (%s, %s, %s, %s)",
-                (search_query, json.dumps(parameters), datetime.now(), jobs_found)
-            )
-            self.conn.commit()
-        
-    def get_all_jobs(self) -> pd.DataFrame:
-        """Get all jobs from the database.
-        
-        Returns:
-            DataFrame containing all jobs
-        """
-        self._ensure_connection()
-        return pd.read_sql("SELECT * FROM scraped_jobs", self.conn)
-    
-    def get_jobs_by_query(self, search_query: str) -> pd.DataFrame:
-        """Get jobs matching a specific search query.
-        
-        Args:
-            search_query: The search query to filter by
-            
-        Returns:
-            DataFrame containing matching jobs
-        """
-        self._ensure_connection()
-        query = "SELECT * FROM scraped_jobs WHERE search_query = %s"
-        return pd.read_sql(query, self.conn, params=(search_query,))
-    
-    def get_duplicate_groups(self) -> List[List[Dict]]:
-        """Identify groups of duplicate jobs based on title and company.
-        
-        Returns:
-            List of groups, where each group is a list of job dictionaries
-        """
-        self._ensure_connection()
-        
-        query = """
-        SELECT id, site, title, company, description, min_amount, max_amount, job_url, is_remote, location, search_query, date_posted
-        FROM scraped_jobs 
-        WHERE title IS NOT NULL AND company IS NOT NULL
-        ORDER BY title, company, site
-        """
-        
-        with self.conn.cursor(cursor_factory=psycopg2.extras.RealDictCursor) as cursor:
-            cursor.execute(query)
-            rows = cursor.fetchall()
-        
-        # Group jobs by (title, company) combination
-        groups = defaultdict(list)
-        for row in rows:
-            job_dict = dict(row)
-            key = (row['title'].strip().lower() if row['title'] else '', 
-                   row['company'].strip().lower() if row['company'] else '')
-            groups[key].append(job_dict)
-        
-        # Filter out groups with only one job (not duplicates)
-        duplicate_groups = [group for group in groups.values() if len(group) > 1]
-        
-        return duplicate_groups
-    
-    def clear_all_jobs(self) -> int:
-        """Clear all data from the scraped_jobs table without deleting the table.
-        
-        Returns:
-            Number of rows deleted
-        """
-        self._ensure_connection()
-        
-        with self.conn.cursor() as cursor:
-            cursor.execute("DELETE FROM scraped_jobs")
-            rows_deleted = cursor.rowcount
-            self.conn.commit()
-            
-        logger.info(f"Cleared {rows_deleted} rows from scraped_jobs table")
-        return rows_deleted
-    
-    def delete_jobs_before_date(self, date_str: str) -> int:
-        """Delete jobs scraped before a specified date.
-        
-        Args:
-            date_str: Date string in YYYY-MM-DD format
-            
-        Returns:
-            Number of rows deleted
-        """
-        self._ensure_connection()
-        
-        try:
-            date_obj = datetime.strptime(date_str, "%Y-%m-%d")
-            
-            with self.conn.cursor() as cursor:
-                cursor.execute(
-                    "DELETE FROM scraped_jobs WHERE date_scraped < %s", 
-                    (date_obj,)
-                )
-                rows_deleted = cursor.rowcount
-                self.conn.commit()
-                
-            logger.info(f"Deleted {rows_deleted} jobs scraped before {date_str}")
-            return rows_deleted
-        except Exception as e:
-            logger.error(f"Error deleting jobs by date: {str(e)}")
-            return 0
-    
-    def delete_jobs_by_ids(self, ids_file: str) -> int:
-        """Delete jobs by their IDs from a file.
-        
-        Args:
-            ids_file: Path to a file containing job IDs, one per line
-            
-        Returns:
-            Number of rows deleted
-        """
-        self._ensure_connection()
-        
-        try:
-            if not os.path.exists(ids_file):
-                logger.warning(f"IDs file {ids_file} not found")
-                return 0
-                
-            with open(ids_file, 'r') as f:
-                job_ids = [line.strip() for line in f if line.strip()]
-                
-            if not job_ids:
-                logger.warning(f"No IDs found in {ids_file}")
-                return 0
-                
-            with self.conn.cursor() as cursor:
-                cursor.execute("DELETE FROM scraped_jobs WHERE id = ANY(%s)", (job_ids,))
-                rows_deleted = cursor.rowcount
-                self.conn.commit()
-                
-            logger.info(f"Deleted {rows_deleted} jobs by ID from {ids_file}")
-            return rows_deleted
-        except Exception as e:
-            logger.error(f"Error deleting jobs by IDs: {str(e)}")
-            return 0
-    
-    def delete_jobs_by_salary(self, min_threshold: int = 70000, max_threshold: int = 90000) -> int:
-        """Delete jobs with salaries below specified thresholds.
-        
-        Args:
-            min_threshold: Minimum salary threshold (default: 70000)
-            max_threshold: Maximum salary threshold (default: 90000)
-            
-        Returns:
-            Number of rows deleted
-        """
-        self._ensure_connection()
-        
-        try:
-            query = """
-            DELETE FROM scraped_jobs
-            WHERE 
-              (
-                min_amount != 0 
-                AND min_amount < %s 
-                AND max_amount < %s
-              )
-              OR 
-              (
-                min_amount >= %s 
-                AND max_amount < %s
-              )
-            """
-            
-            with self.conn.cursor() as cursor:
-                cursor.execute(query, (min_threshold, max_threshold, min_threshold, max_threshold))
-                rows_deleted = cursor.rowcount
-                self.conn.commit()
-                
-            logger.info(f"Deleted {rows_deleted} jobs with salaries below thresholds (min: {min_threshold}, max: {max_threshold})")
-            return rows_deleted
-        except Exception as e:
-            logger.error(f"Error deleting jobs by salary: {str(e)}")
-            return 0
-    
-    def delete_jobs_by_field(self, field: str, patterns_file: str) -> int:
-        """Delete jobs by matching patterns on a specific field.
-        
-        Args:
-            field: The database field to match on (e.g., 'company', 'title')
-            patterns_file: Path to a file containing patterns, one per line
-            
-        Returns:
-            Number of rows deleted
-        """
-        self._ensure_connection()
-        
-        try:
-            if not os.path.exists(patterns_file):
-                logger.warning(f"Patterns file {patterns_file} not found")
-                return 0
-                
-            with open(patterns_file, 'r') as f:
-                patterns = [line.strip() for line in f if line.strip()]
-                
-            if not patterns:
-                logger.warning(f"No patterns found in {patterns_file}")
-                return 0
-                
-            # Validate field name to prevent SQL injection
-            valid_fields = {'company', 'title'}
-            if field not in valid_fields:
-                logger.error(f"Invalid field name: {field}")
-                return 0
-                
-            rows_deleted = 0
-            
-            with self.conn.cursor() as cursor:
-                for pattern in patterns:
-                    # Convert pattern to lowercase for case-insensitive matching
-                    pattern_lower = pattern.lower()
-                    query = sql.SQL("DELETE FROM scraped_jobs WHERE LOWER({}) LIKE %s").format(
-                        sql.Identifier(field)
-                    )
-                    cursor.execute(query, (pattern_lower,))
-                    pattern_deleted = cursor.rowcount
-                    rows_deleted += pattern_deleted
-                    
-                    if pattern_deleted > 0:
-                        logger.info(f"Pattern '{pattern}' deleted {pattern_deleted} rows")
-                    else:
-                        logger.debug(f"Pattern '{pattern}' found no matches")
-                
-                self.conn.commit()
-                
-            logger.info(f"Deleted {rows_deleted} jobs matching {field} patterns from {patterns_file}")
-            return rows_deleted
-        except Exception as e:
-            logger.error(f"Error deleting jobs by {field}: {str(e)}")
-            return 0
-    
-    def backup_and_reset(self) -> bool:
-        """Create a backup of the database and clear all data.
-        
-        Returns:
-            True if operation was successful, False otherwise
-        """
-        try:
-            # Create pg_dump backup first
-            backup_info = self.create_backup('manual', 'backup_and_reset')
-            print(f"✓ Database backup created: {backup_info['filename']} ({backup_info['size_mb']} MB)")
-            
-            # Also create CSV backup for backwards compatibility
-            backup_dir = os.path.join(SCRIPT_DIR, "backups")
-            os.makedirs(backup_dir, exist_ok=True)
-            
-            timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-            
-            # Backup scraped_jobs
-            jobs_df = self.get_all_jobs()
-            jobs_backup_path = os.path.join(backup_dir, f"scraped_jobs_{timestamp}.csv")
-            jobs_df.to_csv(jobs_backup_path, index=False)
-            logger.info(f"Jobs CSV backup created at {jobs_backup_path}")
-            
-            # Backup search_history
-            search_df = pd.read_sql("SELECT * FROM search_history", self.conn)
-            search_backup_path = os.path.join(backup_dir, f"search_history_{timestamp}.csv")
-            search_df.to_csv(search_backup_path, index=False)
-            logger.info(f"Search history CSV backup created at {search_backup_path}")
-            
-            # Clear all data
-            rows_deleted = self.clear_all_jobs()
-            
-            with self.conn.cursor() as cursor:
-                cursor.execute("DELETE FROM search_history")
-                search_rows_deleted = cursor.rowcount
-                self.conn.commit()
-            
-            logger.info(f"Database reset completed. Jobs: {rows_deleted}, Search history: {search_rows_deleted}")
-            
-            return True
-        except Exception as e:
-            logger.error(f"Error during database backup and reset: {str(e)}")
-            return False
-    
-    def close(self) -> None:
-        """Close the database connection."""
-        if self.conn and not self.conn.closed:
-            self.conn.close()
-            logger.info("Database connection closed")
 
 
 class DuplicateManager:
